requirements.txt
Django==1.11
oauth==1.0.1
oauth2==1.9.0post1
oauth2client==4.1.2
oauthlib==2.0.6
python-social-auth[django]==0.3.6
requests==2.18.4
requests-oauth2==0.2.0
requests-oauthlib==0.8.0
pytz==2017.3
django-localflavor==2.0
python_dateutil==2.6.1
